Patch powerlevel10k internal VPN prompt
This is a hack because there's no support for color classes in the VPN
prompt as of this commit.

This commit adds a conditional block to render the prompt in an orangish
color when it's a certain IP (the kind of VPNs you don't want to be using
at all times) and green for all other IPs.

Signed-off-by: PrayagS <prayagsavsani@gmail.com>