feat(neovim): update fold configuration and remove nvim-ufo
Thanks to a recent post on r/neovim[^1], it seems
like nvim-ufo is not really needed anymore.

Now the config is set as follows,
1. set `foldexpr` to treesitter in opts.lua
2. use `vim.lsp.foldexpr` when an LSP server is
attached and the client supports folding.

This commit also removes statuscol.nvim which is
replaced by the statuscolumn module in snacks.nvim.

Post this change and given neovim v0.11, restoring
sessions is broken due to a race condition between
restoring and the LSP server being ready[^2].

As a workaround, the version of auto-session is
replaced with a fork that has the required fix[^3].

[^1]: https://www.reddit.com/r/neovim/comments/1jmqd7t/sorry_ufo_these_7_lines_replaced_you
[^2]: https://github.com/rmagatti/auto-session/issues/409
[^3]: https://github.com/rmagatti/auto-session/issues/409#issuecomment-2764973274

Signed-off-by: PrayagS <prayagsavsani@gmail.com>