fix(wezterm/resurrect): get current mux window _after_ creating the workspace
The `restore_workspace` function takes in the window where we want the
restored workspace to be opened. Default behavior is to open a new
window.

The current window's object was being stored in opts before the new
workspace is created (more details in the comments). This was causing
wezterm to complain about the window being missing when trying to
restore the state. Didn't realize this was happening until I checked the
debug overlay.

Signed-off-by: PrayagS <prayagsavsani@gmail.com>