fix(wezterm): unix domain -> local domain
The wezterm API to get current pane text and inject text into a pane
only works on local domains as of now. Using an unix domain was preventing
resurrect.wezterm from storing the currently running process and all the
text present in a pane.

When using unix domains, only pane layout was being restored by the
plugin. That is no longer the case after this change.

Signed-off-by: PrayagS <prayagsavsani@gmail.com>